📦 CHORE: change exports of Noddles utils
